Bk edit account

+++ Backend-SEP4/Intermediary-DB/Logic/AccountLogic.cs
@@ -1,4 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using System.Security.Cryptography;
+using System.Text;
 using DBComm.Repository;
 using DBComm.Shared;
 
@@ -7,10 +9,22 @@ namespace WebAPI.Service;
 public class AccountLogic : IAccountLogic
 {
     private IAccountRepository _repository;
-    //maybe private
     public AccountLogic(IAccountRepository repository)
     {
         this._repository = repository;
+    }
+
+    private async Task<string> _hashPassword(string password)
+    {
+        byte[] inputBytes = Encoding.UTF8.GetBytes(password);
+        string hashedString = "";
+        using (SHA256 sha256 = SHA256.Create())
+        {
+            byte[] hashBytes = sha256.ComputeHash(inputBytes);
+            hashedString = BitConverter.ToString(hashBytes).Replace("-", "");
+        }
+
+        return hashedString;
     }
     public Task<Member> GetAdmin(string login, string password)
     {
@@ -28,30 +42,38 @@ public class AccountLogic : IAccountLogic
         {
             throw new ValidationException("Password cannot be null");
         }
-        
+
         try
         {
-            return await _repository.RegisterMember(username, password);
-
+            if (await _repository.CheckExistingUser(username))
+            {
+                string hash = await _hashPassword(password);
+                await _repository.RegisterMember(username, hash);
+            }
+        }catch(Exception e)
+        {
+            Console.WriteLine(e);
+            throw new Exception(e.Message);
         }
-        catch(Exception e)
+
+        return null;
+    }
+
+    public async Task Delete(string username,string password)
+    {
+        try
+        {
+            string hash = await _hashPassword(password);
+            if (await _repository.CheckNonExistingUser(username,hash))
+            {
+                await _repository.DeleteAccount(username);
+            }
+        }
+        catch (Exception e)
         {
             throw new Exception(e.Message);
         }
     }
-
-    public async Task Delete(string username)
-    {
-        try
-        {
-            await _repository.DeleteAccount(username);
-        }
-        catch (Exception e)
-        {
-            Console.WriteLine("Error deleting account: " + e.Message);
-        }
-    }
-
 
     public Task<Member> GetMember(string login, string password)
     {
@@ -69,8 +91,81 @@ public class AccountLogic : IAccountLogic
         {
             throw new ValidationException("Password cannot be null");
         }
-        await _repository.RegisterAdmin(username, password);
-        
+        try
+        {
+            if (await _repository.CheckExistingUser(username))
+            {
+                string hash = await _hashPassword(password);
+                await _repository.RegisterAdmin(username, hash);
+            }
+        }
+        catch (Exception e)
+        {
+            Console.WriteLine(e);
+            throw new Exception(e.Message);
+        }
+
         return;
     }
+
+
+
+    public async Task EditUsername(string oldUsername, string newUsername,string password)
+    {
+        try
+        {
+            string hash = await _hashPassword(password);
+            if (await _repository.CheckNonExistingUser(oldUsername,hash))
+            {
+                await _repository.EditUsername(oldUsername, newUsername);
+            }
+        }
+        catch (Exception e)
+        {
+            Console.WriteLine(e.Message);
+            throw new Exception(e.Message);
+        }
+    }
+
+    public async Task EditPassword(string username,string oldPassword, string newPassword)
+    {
+        try
+        {
+            string hash = await _hashPassword(oldPassword);
+            if (await _repository.CheckNonExistingUser(username,hash))
+            {
+                string newHash = await _hashPassword(newPassword);
+                if (newHash == hash)
+                {
+                    throw new Exception("Cannot set same password");
+                }
+                await _repository.EditPassword(username, oldPassword, newHash);
+            }
+        }
+        catch (Exception e)
+        {
+            Console.WriteLine(e.Message);
+            throw new Exception(e.Message);
+        }
+    }
+
+    public async Task ToggleAdmin(string adminUsername,string adminPassword, string username)
+    {
+        try
+        {
+            string hash = await _hashPassword(adminPassword);
+            if (await _repository.CheckIfAdmin(adminUsername,hash,username))
+            {
+                await _repository.ToggleAdmin(username);
+            }
+        }
+        catch (Exception e)
+        {
+            Console.WriteLine(e.Message);
+            throw new Exception(e.Message);
+        }
+    }
 }
+
+
+
